feat: 녹음 시작/정지 및 로컬 파일 저장 구현
- AudioRecorder 클래스 제거 후 MainActivity에 통합 (기능이 단순해 별도 클래스 불필요)
- res/mipmap 디렉토리 생성 (핸드폰 실행을 위한 리소스 대응)
- network_security_config.xml 추가 (서버 연결 허용 설정, 환경에 맞게 IP 설정 필요)
- 녹음 파일 경로 환경 변수 수정 (PC에서도 재생 가능하도록 경로 조정)
- AndroidManifest.xml 수정 (마이크, 저장소, 네트워크 등 권한 추가)
