+++ app/src/main/res/xml/network_security_config.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<network-security-config>
+    <domain-config cleartextTrafficPermitted="true">
+        <domain includeSubdomains="true">192.168.0.26</domain> <!-- 너의 서버 IP로 수정 -->
+    </domain-config>
+</network-security-config>
